requirements.txt
Django==3.0
django-ckeditor==6.0.0
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
djangorestframework==3.10.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.6
django-phonenumber-field==5.0.0
PyJWT==1.7.1
